fix route

+++ migrations/15-create-route.js
@@ -13,7 +13,7 @@ module.exports = {
 	      allowNull: false,
       },
       distance: {
-        type: Sequelize.DOUBLE,
+        type: Sequelize.DECIMAL(4,2),
 	      allowNull: false,
       },
       status: {
